Commit 4 - ongoing
Switched the list to appear once a dive has been clicked on the map, currently showing a list with the dives within a 30km radius from the clicked marker.

This list with different dive names, also has an interactable options, whereas if you click them a graph will show up. Graph is not finished.