Added content below and right of chart
Added a canvas for content below the graph and to the right of the graph.

The idea is that all static content from the dive will be placed under the graph, and all the dynamic content to the right. Havent filled out anything on the right, and started with filling out in the bottom.